% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetAnnotation.R
\name{annotate_metabolite}
\alias{annotate_metabolite}
\title{annotate_metabolite}
\usage{
annotate_metabolite(
  object = NULL,
  ms1_file = "data.csv",
  ms2_file = NULL,
  ms2_type = c("mzML", "mzXML", "mgf", "msp"),
  path = ".",
  lib_db,
  polarity = c("positive", "negative"),
  lib = c("dodd"),
  column = c("hilic", "c18"),
  ce = c("20", "10", "40"),
  adduct_list = c("[M+H]+"),
  mz_tol = 15,
  mz_ppm_thr = 150,
  pf_rt_range = 10,
  tolerance_rt_range = 20,
  is_rt_score = TRUE,
  is_ms2_score = TRUE,
  is_include_precursor = TRUE,
  int_ms2_min_abs = 50,
  int_ms2_min_relative = 0.01,
  mz_tol_combine_ms1_ms2 = 20,
  rt_tol_combine_ms1_ms2 = 20,
  mz_tol_ms2 = 35,
  dp_cutoff = 0.8,
  matched_frag_cutoff = 1,
  direction = c("reverse", "forward"),
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  is_plot_ms2 = TRUE
)
}
\arguments{
\item{ms1_file}{The name of ms1 peak table. Column 1 is "name", Column 2 is "mz" and column is "rt".}

\item{ms2_file}{Default: NULL}

\item{ms2_type}{"mgf", "mzXML", "msp", "cef"}

\item{path}{Default: '.'}

\item{lib}{Default: 'zhuMetLib'}

\item{column}{"hilic", "rp"}

\item{ce}{"10", "20", "30", "35,15", "40", "50". Default: "30"}

\item{mz_tol}{ms1 match. Default: 25 ppm}

\item{tolerance_rt_range}{maxmium rt tolerance. Default: 30 s}

\item{is_rt_score}{logical vector. Default: TRUE}

\item{is_ms2_score}{logical vector. Default: TRUE}

\item{is_include_precursor}{Default: TRUE}

\item{int_ms2_min_abs}{Default: 30}

\item{int_ms2_min_relative}{Default: 0.01}

\item{mz_tol_combine_ms1_ms2}{Default: 25 ppm}

\item{rt_tol_combine_ms1_ms2}{Default: 10 s}

\item{mz_tol_ms2}{Default: 35 ppm}

\item{dp_cutoff}{The cutoff of dot product. Default is 0.8.}

\item{matched_frag_cutoff}{The cutoff of matched fragment number. Default: 1}

\item{direction}{Reserve which direct dot product. 'reverse', 'forward'. Default: 'reverse'}

\item{scoring_approach}{spectral match approach, including 'dp', 'bonanza', 'hybrid', 'gnps'. Default: 'dp'.}

\item{is_plot_ms2}{Output MS2 match plot or not. Default: TRUE}

\item{metdna_version}{'version1', 'version2'. Default: "version1"}

\item{instrument}{The instrument you used to acquire data. "AgilentQTOF", "SciexTripleTOF", "BrukerQTOF", "ThermoOrbitrap", "ThermoExploris". Default: "SciexTripleTOF"}

\item{method_lc}{'Amide12min', 'Amide23min'. Default: 'Amide12min'}

\item{excluded_adduct}{adduct list for exclusion. Default: NULL}

\item{is_rt_calibration}{Default: FALSE}

\item{pf_rt_tol}{penalty-free rt range. Default: 0 s}

\item{pf_ccs_tol}{penalty-free rt range. Default: 0 %}

\item{tolerance_ccs_range}{maxmium rt tolerance. Default: 2 %}

\item{is_filter}{Whether filter candidates whose rt/ccs exceed the maximum tolerance. Default: FALSE}

\item{is_ccs_score}{logical vector. Default: FALSE}

\item{ccs_tol_combine_ms1_ms2}{Only appliable for IM-MS data. Default: NULL}
}
\value{
Return the annotation result.
}
\description{
Annotate peak using MS/MS spectra in in-house database.
}
\author{
Zhiwei Zhou
}
