% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetAnnotation.R
\name{convertSpecAnnotationClass2Table}
\alias{convertSpecAnnotationClass2Table}
\title{convertSpecAnnotationClass2Table}
\usage{
convertSpecAnnotationClass2Table(ms1_data, result_annotation, ...)
}
\description{
convertSpecAnnotationClass2Table
}
