% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunIbdProject.R
\name{create_ibd_object}
\alias{create_ibd_object}
\title{create_ibd_object}
\usage{
create_ibd_object(
  peak_table = "Peak-table.csv",
  merged_worklist = "merged_worklist_c18_pos.xlsx",
  dir_path = ".",
  mode = c("c18_pos", "c18_neg", "hilic_pos", "hilic_neg")
)
}
\arguments{
\item{mode}{c('c18_pos', 'c18_neg', 'hilic_pos', 'hilic_neg')}
}
\description{
create_ibd_object
}
\examples{
\dontrun{
peak_table <- 'Peak-table.csv'
dir_path <- '~/Project/00_IBD_project/Data/20230906_B001_B004_batch_merge/C18_pos/00_raw_data_processing'
merged_worklist <- '~/Project/00_IBD_project/Data/00_IBD_worklists & sample_id/merged_worklist/merged_worklist_c18_pos.xlsx'
mode <- 'c18_pos'
create_ibd_object(peak_table = 'Peak-table.csv',
                  merged_worklist = '~/Project/00_IBD_project/Data/00_IBD_worklists & sample_id/merged_worklist/merged_worklist_c18_pos.xlsx',
                  dir_path = '~/Project/00_IBD_project/Data/20230906_B001_B004_batch_merge/C18_pos/00_raw_data_processing',
                  mode = 'c18_pos')
}
}
\author{
Zhiwei Zhou
}
