% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetAnnotation.R
\name{generate_summary_table}
\alias{generate_summary_table}
\title{generate_summary_table}
\usage{
generate_summary_table(result_annotation)
}
\description{
generate a summary table of metabolite annotation
}
\author{
Zhiwei Zhou
}
