% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunDbSearch.R
\name{get_compound}
\alias{get_compound}
\title{get_compound}
\usage{
get_compound(lab_id)
}
\arguments{
\item{lab_id}{Dodd lab ID}
}
\description{
search compound information via lab_id
}
\examples{
\dontrun{
get_compound(lab_id = "S00001")
}
}
\author{
Zhiwei Zhou
}
