% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunHandlingMassdataset.R
\name{match_biomet}
\alias{match_biomet}
\title{match_biomet}
\usage{
match_biomet(
  object,
  mode = c("hilic_pos", "hilic_neg", "c18_pos", "c18_neg"),
  lib_version = c("v1"),
  mz_ppm = 10,
  rt_second = 15
)
}
\arguments{
\item{mode}{data modes. "hilic_pos", "hilic_neg", "c18_pos", "c18_neg"}

\item{mz_ppm}{mz tolerance (in ppm) to match biomet. defaule: 10 ppm}

\item{rt_second}{rt tolerance (in second) to match biomet. defaule: 25 second}

\item{obj}{a object of mass_dataset}
}
\description{
match some real metabolite in peak table according to the data mode
}
\examples{
\dontrun{
result <- match_istd(object = object,
  mode = 'hilic_pos',
  mz_ppm = 10,
  rt_second = 15)
}
}
\author{
Zhiwei Zhou
}
