% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunHandlingMassdataset.R
\name{match_istd}
\alias{match_istd}
\title{match_istd}
\usage{
match_istd(
  object,
  mode = c("hilic_pos", "hilic_neg", "c18_pos", "c18_neg"),
  lib_version = c("v2", "v1"),
  mz_ppm = 10,
  rt_second = 15
)
}
\arguments{
\item{mode}{data modes. "hilic_pos", "hilic_neg", "c18_pos", "c18_neg"}

\item{mz_ppm}{mz tolerance (in ppm) to match ISTD. defaule: 10 ppm}

\item{rt_second}{rt tolerance (in second) to match ISTD. defaule: 25 second}

\item{obj}{a object of mass_dataset}
}
\description{
match isotope labeling internal standards (ISTD) in peak table according to the data mode
}
\examples{
\dontrun{
result <- match_istd(object = object,
  mode = 'hilic_pos',
  mz_ppm = 10,
  rt_second = 15)
}
}
\author{
Zhiwei Zhou
}
