% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetAnnotation.R
\name{match_ms1_rt}
\alias{match_ms1_rt}
\title{match_ms1_rt}
\usage{
match_ms1_rt(
  ms1_data,
  lib_meta,
  mz_tol = 15,
  mz_ppm_thr = 150,
  pf_rt_range = 0,
  tolerance_rt_range = 20,
  is_combined_rt = FALSE,
  is_rt_score = FALSE,
  is_check_cation = FALSE
)
}
\arguments{
\item{ms1_data}{ms1 data, including info and subject.}

\item{lib_meta}{the meta information of library}

\item{mz_tol}{m/z tolerance. Default: 10ppm}

\item{pf_rt_range}{Default: 0}

\item{tolerance_rt_range}{Defaule: 20}

\item{is_combined_rt}{Defaule: FALSE}

\item{is_rt_score}{Defaule: FALSE}
}
\description{
match_ms1_rt
}
\author{
Zhiwei Zhou
}
