% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunIbdProject.R
\name{merge_worklist_sample_info}
\alias{merge_worklist_sample_info}
\title{merge_worklist_sample_info}
\usage{
merge_worklist_sample_info(
  sample_info_files,
  worklist_files,
  path = ".",
  mode = c("c18_pos", "c18_neg", "hilic_pos", "hilic_neg")
)
}
\description{
modify the worklist for data processing (LC-MS worklist)
}
\examples{
\dontrun{
setwd('~/Project/00_IBD_project/Data/00_IBD_worklists & sample_id/')
file_name <- '~/Project/00_IBD_project/Data/00_IBD_worklists & sample_id/worklists/B003_Worklist_c18_neg.xlsx'
modify_worklist(file_name)
}
}
\author{
Zhiwei Zhou
}
