% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunIbdProject.R
\name{modify_sample_info}
\alias{modify_sample_info}
\title{modify_sample_info}
\usage{
modify_sample_info(file_name)
}
\description{
modify the sample information for data processing (Aedan extraction table)
}
\examples{
\dontrun{
setwd('~/Project/00_IBD_project/Data/00_IBD_worklists & sample_id/')
file_name <- '~/Project/00_IBD_project/Data/00_IBD_worklists & sample_id/sampleinfo/2023-08-25 IBD enrollment, NewBatch_004 samples.csv'
modify_sample_info(file_name)
}
}
\author{
Zhiwei Zhou
}
