% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunIbdProject.R
\name{modify_worklist}
\alias{modify_worklist}
\title{modify_worklist}
\usage{
modify_worklist(file_name)
}
\description{
modify the worklist for data processing (LC-MS worklist)
}
\examples{
\dontrun{
setwd('~/Project/00_IBD_project/Data/00_IBD_worklists & sample_id/')
file_name <- '~/Project/00_IBD_project/Data/00_IBD_worklists & sample_id/worklists/B003_Worklist_c18_neg.xlsx'
modify_worklist(file_name)
}
}
\author{
Zhiwei Zhou
}
