% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunPlots.R
\name{plot_id_shift_ms2}
\alias{plot_id_shift_ms2}
\title{plot_id_shift_ms2}
\usage{
plot_id_shift_ms2(obj_ms2 = NULL, obj_spec_frag_match, obj_spec_frag_nl)
}
\arguments{
\item{obj_ms2}{object of SpectraData}
}
\description{
generate ms2 mirror plot for shift match (bonanza, hybrid, gnps)
}
\author{
Zhiwei Zhou
}
