% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunPlots.R
\name{plot_ms2}
\alias{plot_ms2}
\title{plot_ms2}
\usage{
plot_ms2(ms2_spec, is_normalize = TRUE)
}
\arguments{
\item{ms2_spec}{a table of spectrum. 1st column "mz", 2nd column "intensity"}

\item{is_normalize}{whether normalize the spectrum}
}
\description{
plot ms2 spec
}
\examples{
\dontrun{
ms2_spec <- get_ms2(lab_id = 'S00001', ce = '20', polarity = 'positive')
plot_ms2(ms2_spec, is_normalize = FALSE)
plot_ms2(ms2_spec, is_normalize = TRUE)
}
}
\author{
Zhiwei Zhou
}
