% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetAnnotation.R
\name{read_ms1}
\alias{read_ms1}
\title{read_ms1}
\usage{
read_ms1(filename, path = ".")
}
\arguments{
\item{filename}{The name of csv file.}

\item{path}{'.'}

\item{instrument}{"SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap", 'ThermoExploris', "AgilentDTIMMS", "BrukerTIMS", 'WatersQTOF','WatersTWIMMS'}
}
\value{
A list. info: a table of feature information, column 1_feature name, column 2_mz, column3_rt in seconds, column4_ccs. subject: a table of peak area.
}
\description{
read_ms1
}
\author{
Zhiwei Zhou
}
