% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunDbSearch.R
\name{run_spec_match}
\alias{run_spec_match}
\title{run_spec_match}
\usage{
run_spec_match(
  obj_ms2_cpd1,
  obj_ms2_cpd2,
  mz_tol_ms2 = 35,
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  ...
)
}
\arguments{
\item{obj_ms2_cpd1}{experimental ms2 object. Note: info is a data.frame, spec is a list formed by matrix}

\item{obj_ms2_cpd2}{library ms2 object.}

\item{mz_tol_ms2}{Default: 35 ppm}

\item{scoring_approach}{'dp', 'bonanza', 'hybrid', 'gnps'}
}
\description{
a wrapper of runing SpectraTools
}
\author{
Zhiwei Zhou
}
